requirements.txt
aiofiles==0.6.0
aiohttp==3.8.3
async-timeout==4.0.2
#attrs==20.2.0
bech32==1.2.0
#chardet==3.0.4
discord==2.0.0
discord.py==2.0.1
wheel==0.36.2
#idna==2.10
#idna-ssl==1.1.0
#mnemonic
#multidict==6.0.2
#pkg-resources
#six==1.15.0
#typing-extensions==3.7.4.3
#yarl==1.8.1
tabulate==0.8.10
mospy-wallet==0.3.3
python-dotenv==0.21.0
evmospy==1.0.2
charset-normalizer
web3-ethereum-defi[data]
web3